+++ src/opentaskpy/remotehandlers/local.py
@@ -3,7 +3,6 @@
 This module contains the local handlers for transfers.
 """
 
-import glob
 import logging
 import os
 import random
@@ -121,7 +120,7 @@ class LocalTransfer(RemoteTransferHandler):
         """
         return 0
 
-    def push_files_from_worker(self, local_staging_directory: str) -> int:
+    def push_files_from_worker(self, local_staging_directory: str, files: list) -> int:
         """Push files from the worker to another local directory.
 
         Moves files from a local staging directory, to another local destination folder.
@@ -129,6 +128,7 @@ class LocalTransfer(RemoteTransferHandler):
         Args:
             local_staging_directory (str): The local staging directory to upload the
             files from.
+            files (list): A list of files to upload.
 
         Returns:
             int: 0 if successful, 1 if not.
@@ -163,8 +163,7 @@ class LocalTransfer(RemoteTransferHandler):
 
         # Transfer the files
         result = 0
-        # Get list of files in local_staging_directory
-        files = glob.glob(f"{local_staging_directory}/*")
+
         for file in files:
             file_name = os.path.basename(file)
             final_destination = f"{destination_directory}/{file_name}"